requirements.txt
Flask
Flask-WTF
WTForms
Flask-MySQLdb
mysqlclient
Flask-Bootstrap
Flask-Login
email-validator
python-dotenv